rename widget for clarity
